tests/wpt/web-platform-tests/largest-contentful-paint/image-not-fully-visible.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: size when image overflows</title>
<!-- In this test, an image with an intrinsic size of 100 x 50 is added, but
  scaled up in order to overflow the viewport. -->
<body>
<style>
body {
  /* Preventing a scrollbar from showing and removing any margins simplifies
     the calculations below. */
  overflow: hidden;
  margin: 0px;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<script>
  let beforeRender;
  const viewportWidth = document.documentElement.clientWidth;
  const viewportHeight = document.documentElement.clientHeight;
  async_test(function (t) {
    if (!window.PerformanceElementTiming) {
      assert_unreached("PerformanceElementTiming is not implemented");
    }
    const beforeLoad = performance.now();
    new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        const url = window.location.origin + '/images/green-100x50.png';
        // To compute the size, compute the percentage of the image visible and
        // scale by its natural dimensions. In this test, the image is twice
        // the viewport width and twice the viewport height, so it is
        // effectively clipped to 50% by 50% of its display size. Scaling by
        // its natural width and height of 100px and 50px respectively, leads
        // to a weighted size of 50 by 25.
        const truncatedWidth = 100 / 2;
        const truncatedHeight = 50 / 2;
        const weightedSize = truncatedWidth * truncatedHeight;
        checkImage(entry, url, 'image_id', weightedSize, beforeLoad);
      })
    ).observe({type: 'largest-contentful-paint', buffered: true});
    // Add an image, setting width and height equal to viewport.
    img = document.createElement('img');
    img.src = '/images/green-100x50.png';
    img.id = 'image_id';
    img.width = viewportWidth * 2;
    img.height = viewportHeight * 2;
    document.body.appendChild(img);
  }, 'The intersectionRect of an img element overflowing is computed correctly');
</script>
</body>
